+++ tests/test_raillabel/format/test_transform.py
@@ -24,7 +24,7 @@ def transform_json(point3d_dict, quaternion_dict) -> JSONTransformData:
 
 @pytest.fixture
 def transform(point3d, quaternion) -> Transform:
-    return Transform(pos=point3d, quat=quaternion)
+    return Transform(position=point3d, quaternion=quaternion)
 
 
 # == Tests ============================
@@ -38,12 +38,12 @@ def test_from_json(transform_json, transform):
 def test_fromdict(point3d, point3d_dict, quaternion, quaternion_dict):
     transform = Transform.fromdict({"translation": point3d_dict, "quaternion": quaternion_dict})
 
-    assert transform.pos == point3d
-    assert transform.quat == quaternion
+    assert transform.position == point3d
+    assert transform.quaternion == quaternion
 
 
 def test_asdict(point3d, point3d_dict, quaternion, quaternion_dict):
-    transform = Transform(pos=point3d, quat=quaternion)
+    transform = Transform(position=point3d, quaternion=quaternion)
 
     assert transform.asdict() == {"translation": point3d_dict, "quaternion": quaternion_dict}